Separated the prototype into plugin and test apps.
The plugin implementation has moved to the gradle/ directory,
and the test applications have moved under the testapps/
directory. This change means that you now have to run
./gradlew uploadArchives in the root directory to make the
plugin available to the test application builds via a
local repository (in repo/).

See the readme.md for more details about building and using
the test applications.

Change-Id: Id9b268aa1d6d1a5621513fb788b2dc90ca2793e8