Add support to build and bundle library projects.
The unarchived bundle is saved under build/bundle.

The generated manifest is directly saved under there
to avoid copying it.
The resources are copied there but the built-type
res folder is not properly merged in there. Need
to fix this.
The crunched resources are copied on top of it.

Also add a library test app.

Change-Id: Iab2d437b6f98ca5f842ca9720aafa206691fc1f0