Fix support for building and testing libraries.
Libraries weren't properly builds due to:
--non-constant-id was not used
--extra-packages was not used.

Also this creates a fake AndroidDependency for the tested
config so that it can be added to the dependency of the test
app.

Finally, this fixes some of the issues around classpath.
The Dex task receives separately the compiled code and the
library to planned for pre-dexing the libraries. The libraries
are handled through the AndroidDependency class provided
to the VariantConfig.
The compileTask setup is more precise in what is the
compiled *and packaged* classpath vs just a compile classpath
(this is different for test apps depending on if the test target
is a library or an app).

Change-Id: I0be9da2ce03bb48f40f4577e4323edf50062f2c0