Make sure the build detects changes in config classes.
aapt and dex option classes use the @Input and the
tasks using them put @Nested on them.

The ProductFlavor and the Build Type classes are now
Serializable and they are passed as @Input args to
all classes that need them.

Any change in those objects will properly trigger
a recompilation.

Change-Id: I42fb3f6c5870505b984ff777befee255f03ebd1d