Generate smaller R classes for libraries.
Using the new --output-text-symbols from aapt the build
system now generates the R class for libraries manually
based on the symbols exported by the libraries and the
final values computed by aapt when using all the resource
folders.

Because only R.java is concerned, the Manifest class is
now included in the library jar file.

Also added a new test apps that uses instrumentation to
verify the build system.

Change-Id: I0e9ba124cfba729f4ab8021b440b1ac70c5d1b80