Rework the way tests are installed and run.
It used to do:
    Install all, run all, uninstall all.
It now does:
   foreach variant: install, run, uninstall.

This allows to properly test different flavors that share the same
package name. Before installAll would override some packages with new
ones and only the last installed flavor would be installed.

Note that right now this relies on dependsOn to do the install, run
and uninstall in the proper order which is not reliable so this
needs to be fixed.

Also, we'll want to change this later to support deploying to
multiple emulators.

Change-Id: I0931b3a9b7a0cbd37c2e763e4abf768777342b37