Prepare to upload the gradle plugin to maven.
Change the groupId to com.android.tools.build
Create javadoc and source artifacts.

Move all samples to using the local maven repo,
so that the uploadArchives task can target the
real Maven repo once we're ready to upload there.

Create a release.gradle init script that will
not use the -SNAPSHOT qualifier when building the
artifacts.

Change-Id: Ie17e6299ae5335535c440ab99695e6e408c1f352