Fix manifest merging.
minSdkVersion/targetSdkVersion/versionCode/versionName are now
inserted into the generated/merged manifest instead of when
packaging the APK. This allows proper conflict detection
when merging the manifest of the library projects.

Also move all testapps to not use mavenLocal due to
an issue with it (failure to read resource file inside
the plugin).

Change-Id: I46aa9563cbc31fe45bc0f1a44ed48ae110cb376e