Support compile config for flavors and build types.
Read dependency graph and create list of JarDependency
and AndroidDependency for each configuration.

Set Jars and AndroidDependency on each VariantConfig
object.

Add a AndroidDependency report task (called androidDependency)

The prepareDependenciesTask will prepare the same artifact
multiple time if it's used by different variants and they
are all built. This will be fixed in the next patch.

Change-Id: I4f11565c3629a54f022ecc82f81d4900ef2bd0c9