Per-lib prepare task.
each Android library dependency is assigned its own
task to ensure the library is unzipped.
This task is set to depend on the build dependencies
of the configuration(s) that brought it in the dependency
graph.

The Prepare task for each variant is set to depend
on all the prepareLibTask needed by the dependency
graph.

Change-Id: I17829eac949c12df706d2328441fc90f9c64701d